refactor(auth): update import paths for ErrorResponse type to use the new response.ts file
